pyproject.toml
[project]
name = "osira"
version = "0.1.0"
description = "AI-powered financial intelligence agent"
requires-python = ">=3.11"
dependencies = [
    "redis>=5.0",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["ruff", "pytest"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
